feat: update template based on discussed definitions/settings, GC-3909
