+++ .gitignore
@@ -25,8 +25,8 @@
 !.devcontainer/**
 !.github/
 !.github/**
-!src/
-!src/**
+!python_app_template/
+!python_app_template/**
 !tests/
 !tests/**
 !docs/
