+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
   "typer>=0.15.1",
   "prometheus_client>=0.11.0",
   "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
+  "loguru (>=0.7.3,<0.8.0)",
 ]
 
 [project.scripts]  # https://docs.astral.sh/uv/concepts/projects/config/#command-line-interfaces
@@ -65,13 +66,13 @@ skip_covered = true
 branch = true
 command_line = "--module pytest"
 data_file = "reports/.coverage"
-source = ["src"]
+source = ["python_app_template"]
 
 [tool.coverage.xml]  # https://coverage.readthedocs.io/en/latest/config.html#xml
 output = "reports/coverage.xml"
 
 [tool.basedpyright]  # https://docs.basedpyright.com/v1.28.1/configuration/config-files/
-include = ["src"]
+include = ["python_app_template"]
 reportMissingImports = "error"
 reportMissingTypeStubs = false
 typeCheckingMode = "strict"
@@ -82,23 +83,18 @@ pythonPlatform = "Linux"
 [tool.pytest.ini_options]  # https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
 addopts = "--color=yes --doctest-modules --exitfirst --failed-first --strict-config --strict-markers --verbosity=2 --junitxml=reports/pytest.xml"
 filterwarnings = ["error", "ignore::DeprecationWarning"]
-testpaths = ["src", "tests"]
+testpaths = ["python_app_template", "tests"]
 xfail_strict = true
 
 [tool.ruff]  # https://docs.astral.sh/ruff/settings/
-fix = true
 line-length = 120
-src = ["src", "tests"]
+src = ["python_app_template", "tests"]
 target-version = "py312"
 
 [tool.ruff.format]
 docstring-code-format = true
 
 [tool.ruff.lint]
-select = ["A", "ASYNC", "B", "BLE", "C4", "C90", "D", "DTZ", "E", "EM", "ERA", "F", "FBT", "FLY", "FURB", "G", "I", "ICN", "INP", "INT", "ISC", "LOG", "N", "NPY", "PERF", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "Q", "RET", "RSE", "RUF", "S", "SIM", "SLF", "SLOT", "T10", "T20", "TC", "TID", "TRY", "UP", "W", "YTT"]
-ignore = ["D203", "D213", "E501", "RET504", "S101", "S307", "LOG015"]
-unfixable = ["ERA001", "F401", "F841", "T201", "T203"]
-extend-fixable = ["ERA001", "F401", "F841", "T201", "T203"]
 
 [tool.ruff.lint.flake8-tidy-imports]
 ban-relative-imports = "all"
