--- src/python_app_template/api.py
@@ -1,35 +1,0 @@
-"""Python App Template REST API."""
-
-import asyncio
-import logging
-from collections.abc import AsyncGenerator
-from contextlib import asynccontextmanager
-
-from fastapi import FastAPI
-from prometheus_fastapi_instrumentator import Instrumentator
-
-
-@asynccontextmanager
-async def lifespan(app: FastAPI) -> AsyncGenerator[None]:
-    """Handle FastAPI startup and shutdown events."""
-    # Startup events.
-    for handler in logging.root.handlers:
-        logging.root.removeHandler(handler)
-    yield
-    # Shutdown events.
-
-
-app = FastAPI(lifespan=lifespan)
-
-Instrumentator().instrument(app).expose(app)
-
-
-@app.get("/compute")
-async def compute(n: int = 42) -> int:
-    """Compute the result of a CPU-bound function."""
-
-    def fibonacci(n: int) -> int:
-        return n if n <= 1 else fibonacci(n - 1) + fibonacci(n - 2)
-
-    result = await asyncio.to_thread(fibonacci, n)
-    return result
